requirements.txt
irc
requests
python-dotenv
certifi